+++ controller/users.js
@@ -1,5 +1,5 @@
 var express = require('express');
-const {check, validationResult} = require('express-validator/check');
+const { check, validationResult } = require('express-validator/check');
 var format = require('string-format');
 var languageService = require('../validator/language');
 var logger = require('../util/logger.js');
@@ -16,7 +16,9 @@ var router = express.Router();
 var email;
 var servicePro;
 var salesRep;
-function verifyToken(token, res, lang) {
+var id;
+async function verifyToken(token, res, lang) {
+    id = undefined;
     if (!token) {
         languageService.get_lang(lang, 'NO_TOKEN').then(msg => {
             res.send({
@@ -45,6 +47,19 @@ function verifyToken(token, res, lang) {
         salesRep = decoded.salesRep;
         servicePro = decoded.servicePro;
         email = decoded.email;
+        id = decoded.id;
+        if (!id) {
+            languageService.get_lang(lang, 'FAILED_AUTHENTICATE_TOKEN').then(msg => {
+                res.send({
+                    status: statics.STATUS_FAILURE,
+                    code: codes.TOKEN_MISSING,
+                    message: msg.message,
+                    auth: false,
+                    data: null
+                });
+            });
+            return
+        }
         return decoded.email;
 
     });
@@ -102,14 +117,17 @@ router.post('/create', [
     }
 });
 
-router.put('/edit_profile', function (req, res) {
+router.put('/edit_profile', async function (req, res) {
     lang = req.headers.language;
     var errors = validationResult(req);
     if (errors.array().length == 0) {
         var credentials = req.body;
         var lang = req.headers.language;
         var token = req.headers.authorization;
-        verifyToken(token, res, lang);
+        await verifyToken(token, res, lang);
+        if (!id) {
+            return;
+        }
         if (email) {
             if (credentials.first_name == '') {
                 languageService.get_lang(lang, 'EMPTY_FIELD_FIRST').then(msg => {
@@ -180,11 +198,13 @@ router.put('/edit_profile', function (req, res) {
 }
 );
 // get profile info
-router.get('/profile', function (req, res) {
+router.get('/profile', async function (req, res) {
     var lang = req.headers.language;
     var token = req.headers.authorization;
-    console.log(lang);
-     verifyToken(token, res, lang);
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
     if (email) {
         UserService.GetUser(email).then(user => {
             var errors = validationResult(req);
@@ -223,11 +243,13 @@ router.get('/profile', function (req, res) {
     }
 });
 // get profile info
-router.delete('/delete/:id', function (req, res) {
+router.delete('/delete/:id', async function (req, res) {
     var lang = req.headers.language;
     var token = req.headers.authorization;
-
-    verifyToken(token, res, lang);
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
 
     UserService.DeleteUser(req.params.id).then(user => {
         var errors = validationResult(req);
@@ -265,11 +287,14 @@ router.delete('/delete/:id', function (req, res) {
     });
 });
 // get profile info
-router.get('/details/:id', function (req, res) {
+router.get('/details/:id', async function (req, res) {
     var lang = req.headers.language;
-    var token = req.headers.authorization;
 
-    verifyToken(token, res, lang);
+    var token = req.headers.authorization;
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
 
     UserService.GetAllUserData(req.params.id).then(user => {
         var errors = validationResult(req);
@@ -307,17 +332,19 @@ router.get('/details/:id', function (req, res) {
     });
 });
 // admin update user user_type
-router.put('/admin/user_type', function (req, res) {
+router.put('/admin/user_type', async function (req, res) {
+
     var lang = req.headers.language;
-    var token = req.headers.authorization;
     var credentials = req.body;
     var errors = validationResult(req);
 
     if (errors.array().length == 0) {
-        verifyToken(token, res, lang);
 
-
-
+        var token = req.headers.authorization;
+        await verifyToken(token, res, lang);
+        if (!id) {
+            return;
+        }
 
         return new Promise(function (resolve, reject) {
             UserService.UpdateUserType(credentials).then(user => {
@@ -348,14 +375,18 @@ router.put('/admin/user_type', function (req, res) {
     }
 });
 // admin update user account_status
-router.put('/admin/account_status', function (req, res) {
+router.put('/admin/account_status', async function (req, res) {
     var lang = req.headers.language;
-    var token = req.headers.authorization;
     var credentials = req.body;
     var errors = validationResult(req);
 
     if (errors.array().length == 0) {
-        verifyToken(token, res, lang);
+        var token = req.headers.authorization;
+        await verifyToken(token, res, lang);
+        if (!id) {
+            return;
+        }
+
         return new Promise(function (resolve, reject) {
             UserService.UpdateAccountStatus(credentials).then(user => {
                 resolve(user);
@@ -385,11 +416,14 @@ router.put('/admin/account_status', function (req, res) {
     }
 });
 // get service provider Full profile
-router.get('/servicePro/profile', function (req, res) {
+router.get('/servicePro/profile', async function (req, res) {
     var lang = req.headers.language;
     var errors = validationResult(req);
     var token = req.headers.authorization;
-    verifyToken(token, res, lang);
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
     if (errors.array().length == 0 && salesRep) {
         return new Promise(function (resolve, reject) {
             UserService.GetUser(email).then(user => {
@@ -420,14 +454,19 @@ router.get('/servicePro/profile', function (req, res) {
         });
     }
 });
-router.put('/edit_profile', function (req, res) {
+router.put('/edit_profile', async function (req, res) {
     lang = req.headers.language;
     var errors = validationResult(req);
     if (errors.array().length == 0) {
         var credentials = req.body;
         var lang = req.headers.language;
+
         var token = req.headers.authorization;
-        verifyToken(token, res, lang);
+        await verifyToken(token, res, lang);
+        if (!id) {
+            return;
+        }
+
         if (email) {
             if (credentials.first_name == '') {
                 languageService.get_lang(lang, 'EMPTY_FIELD_FIRST').then(msg => {
@@ -496,15 +535,15 @@ router.put('/edit_profile', function (req, res) {
     }
 
 });
-
-
-
 // get profile info
-router.delete('/delete/:id', function (req, res) {
+router.delete('/delete/:id', async function (req, res) {
     var lang = req.headers.language;
-    var token = req.headers.authorization;
 
-    verifyToken(token, res, lang);
+    var token = req.headers.authorization;
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
 
     UserService.DeleteUser(req.params.id).then(user => {
         var errors = validationResult(req);
@@ -540,15 +579,16 @@ router.delete('/delete/:id', function (req, res) {
         }
 
     });
-})
-
-
+});
 // get profile info
-router.get('/details/:id', function (req, res) {
+router.get('/details/:id', async function (req, res) {
     var lang = req.headers.language;
-    var token = req.headers.authorization;
 
-    verifyToken(token, res, lang);
+    var token = req.headers.authorization;
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
 
     UserService.GetAllUserData(req.params.id).then(user => {
         var errors = validationResult(req);
@@ -584,20 +624,20 @@ router.get('/details/:id', function (req, res) {
         }
 
     });
-})
+});
+router.put('/admin/user_type', async function (req, res) {
 
-router.put('/admin/user_type', function (req, res) {
     var lang = req.headers.language;
-    var token = req.headers.authorization;
     var credentials = req.body;
     var errors = validationResult(req);
 
     if (errors.array().length == 0) {
-        verifyToken(token, res, lang);
 
-
-
-
+        var token = req.headers.authorization;
+        await verifyToken(token, res, lang);
+        if (!id) {
+            return;
+        }
         return new Promise(function (resolve, reject) {
             UserService.UpdateUserStatus(credentials).then(user => {
                 resolve(user);
@@ -625,7 +665,102 @@ router.put('/admin/user_type', function (req, res) {
             data: errors.array()
         });
     }
-})
+});
 
+
+
+
+
+// get activities for by own user
+router.get('/user/activities', async function (req, res) {
+    var lang = req.headers.language;
+
+    var token = req.headers.authorization;
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
+
+    UserService.getUserActivities(id).then(user => {
+        var errors = validationResult(req);
+        if (errors.array().length == 0) {
+            if (user == null) {
+                languageService.get_lang(lang, 'DATA_NOT_FOUND').then(msg => {
+                    res.json({
+                        status: statics.STATUS_FAILURE,
+                        code: codes.FAILURE,
+                        message: msg.message,
+                        data: user
+                    });
+                });
+            } else {
+                languageService.get_lang(lang, 'DATA_FOUND').then(msg => {
+                    res.json({
+                        status: statics.STATUS_SUCCESS,
+                        code: codes.SUCCESS,
+                        message: msg.message,
+                        data: user
+                    });
+                });
+            }
+        } else {
+            languageService.get_lang(lang, 'INVALID_DATA').then(msg => {
+                res.json({
+                    status: statics.STATUS_FAILURE,
+                    code: codes.INVALID_DATA,
+                    message: msg.message,
+                    data: errors.array()
+                });
+            })
+        }
+
+    });
+});
+
+// get activities for by admin
+router.get('/admin/activities/:user_id', async function (req, res) {
+    var lang = req.headers.language;
+
+    var token = req.headers.authorization;
+    await verifyToken(token, res, lang);
+    if (!id) {
+        return;
+    }
+
+    UserService.getUserActivities(req.params.user_id).then(user => {
+        var errors = validationResult(req);
+        if (errors.array().length == 0) {
+            if (user == null) {
+                languageService.get_lang(lang, 'DATA_NOT_FOUND').then(msg => {
+                    res.json({
+                        status: statics.STATUS_FAILURE,
+                        code: codes.FAILURE,
+                        message: msg.message,
+                        data: user
+                    });
+                });
+            } else {
+                languageService.get_lang(lang, 'DATA_FOUND').then(msg => {
+                    res.json({
+                        status: statics.STATUS_SUCCESS,
+                        code: codes.SUCCESS,
+                        message: msg.message,
+                        data: user
+                    });
+                });
+            }
+        } else {
+            languageService.get_lang(lang, 'INVALID_DATA').then(msg => {
+                res.json({
+                    status: statics.STATUS_FAILURE,
+                    code: codes.INVALID_DATA,
+                    message: msg.message,
+                    data: errors.array()
+                });
+            })
+        }
+
+    });
+});
 
 module.exports = router;
